# documentation/swagger.yaml
openapi: 3.0.0
info:
  title: MamaCheck API - Complete Reference
  description: |
    # MamaCheck - SMS-First Antenatal Care System

    Comprehensive REST API for pregnancy management, ANC reminders, and danger sign triage in Nigeria.

    ## Key Features
    - **Pregnancy Management**: Registration with OTP verification, timeline tracking
    - **ANC Reminders**: Daily scheduling at 07:00 WAT, personalized messages
    - **Danger Sign Triage**: Weekly check-ins with RED/YELLOW/GREEN outcomes
    - **CHEW Dashboard**: Real-time KPIs, women registry, red flag management
    - **Multi-language**: English, Pidgin, Yoruba, Hausa, Igbo
    - **SMS Integration**: Termii SMS gateway with DND compliance
    - **Visit Tracking**: Attendance marking with 10-minute undo window
    - **Reference Data**: LGA and PHC management with geolocation

    ## Security
    - JWT authentication for all protected endpoints
    - Role-based access (PATIENT, CHEW, SUPERVISOR, ADMIN)
    - OTP verification for phone number validation
    - Rate limiting on SMS and registration endpoints
    - BCRYPT password hashing in production

    ## SMS Triage Guide
    Women reply with symptom numbers:
    - **0** = No symptoms (GREEN)
    - **1** = Heavy bleeding (RED)
    - **2** = Severe headache (RED)
    - **3** = Swollen face/hands (RED)
    - **4** = Blurry vision (YELLOW)
    - **5** = Fever (YELLOW)
    - **6** = Reduced baby movement (YELLOW)
...
  version: "1.0.0"
...
servers:
...
paths:
...
  /webhook/termii/sms:
    post:
      tags:
        - Webhooks
      summary: Incoming SMS Webhook (Termii)
      description: |
        Receives SMS from Termii for:
        - Danger sign triage responses (numbered symptoms)
        - STOP keyword for SMS opt-out
        - Emergency contact information
      operationId: handleIncomingSMS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - from
                - text
              properties:
                from:
                  type: string
                  example: "09012345678"
                text:
                  type: string
                  example: "1,2,3"
                message_id:
                  type: string
      responses:
        "200":
          description: SMS processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: string
                    enum: [processed, opt_out_processed, ignored]
                  triage:
                    type: string
                    enum: [GREEN, YELLOW, RED]
                    description: Triage outcome
                  reportId:
                    type: string
        "404":
          description: User or pregnancy not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...